autoExec 1.0 发布版

+++ src/main/java/com/magic/ssh/controller/ActionController.java
@@ -1,7 +1,9 @@
 package com.magic.ssh.controller;
 
 import com.magic.ssh.entity.Action;
+import com.magic.ssh.entity.Task;
 import com.magic.ssh.service.ActionService;
+import com.magic.ssh.service.TaskService;
 import com.magic.ssh.util.Result;
 import com.magic.ssh.util.ResultCode;
 import org.springframework.validation.annotation.Validated;
@@ -10,22 +12,24 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Objects;
 
-@CrossOrigin
 @RestController
 @RequestMapping("/api/action")
 public class ActionController {
 
     private final ActionService actionService;
 
-    public ActionController(ActionService actionService) {
+    private final TaskService taskService;
+
+    public ActionController(ActionService actionService, TaskService taskService) {
         this.actionService = actionService;
+        this.taskService = taskService;
     }
 
     @GetMapping("/query")
     public Result guery(@RequestParam Integer actionId) {
         Action action = actionService.getActionInfoById(actionId);
         if (Objects.isNull(action)) {
-            return Result.build(ResultCode.OP_ERROR);
+            return Result.build(ResultCode.RESOURCE_NOT_EXIST);
         } else {
             return Result.success(action);
         }
@@ -33,13 +37,7 @@ public class ActionController {
 
     @GetMapping("/queryByUser")
     public Result gueryRoles(@RequestParam Integer userId) {
-
-        List<Action> list = actionService.getActionByUser(userId);
-        if (list.isEmpty()) {
-            return Result.build(ResultCode.OP_ERROR);
-        } else {
-            return Result.success(list);
-        }
+        return  Result.success(actionService.getActionByUser(userId));
     }
 
     @PostMapping("/add")
@@ -64,6 +62,13 @@ public class ActionController {
 
     @PostMapping("/delete")
     public Result delete(@Validated(Action.Update.class) @RequestBody Action action) {
+
+        Task task = taskService.getTaskByAction(action.getActionId());
+        if (Objects.nonNull(task)) {
+            return Result.build(ResultCode.OP_ERROR, "存在关联的任务 " + task.getTaskName() + ",无法删除该动作",
+                    "");
+        }
+
         if (actionService.deleteAction(action.getActionId()) > 0) {
             return Result.success();
         } else {
